+++ scout/constants/variants_export.py
@@ -1,0 +1,10 @@
+EXPORT_HEADER = [
+    'Rank_score',
+    'Chromosome',
+    'Position',
+    'Change',
+    'Position_Change',
+    'HGNC_id',
+    'Gene_name',
+    'Canonical_transcript_HGVS'
+]