+++ CHANGELOG.md
@@ -37,6 +37,7 @@ $scout update hpo
 - Fix problem with parsing gnomad from VEP
 
 ### Added
+- Added a "Filter and export" button to export custom-filtered SNVs to CSV file
 - Dismiss SVs
 - Added IGV alignments viewer
 - Read delivery report path from case config or CLI command